Fix not found ShortcodeTest class by the autoloader
